updated CMakeFile
removed outdated imports

moved _get_backend import _base.py
