+++ onedal/basic_statistics/basic_statistics.py
@@ -22,7 +22,6 @@ import numpy as np
 from onedal import _backend
 
 from ..common._base import BaseEstimator
-from ..common._policy import _get_policy
 from ..datatypes import _convert_to_supported, from_table, to_table
 
 
@@ -46,9 +45,6 @@ class BaseBasicStatistics(metaclass=ABCMeta):
             "sum_squares_centered",
             "second_order_raw_moment",
         ]
-
-    def _get_policy(self, queue, *data):
-        return _get_policy(queue, *data)
 
     def _get_result_options(self, options):
         if options == "all":
