+++ scripts/CMakeLists.txt
@@ -67,7 +67,19 @@ if(ONEDAL_DPCTL_INTEGRATION)
     message(STATUS "DPCTL_INCLUDE_DIR:" ${DPCTL_INCLUDE_DIR})
 endif()
 
+if(IFACE STREQUAL "dpc")
+    set(IFACE_IS_DPC ON)
+else()
+    set(IFACE_IS_DPC OFF)
+endif()
+
 if(IFACE STREQUAL "spmd_dpc")
+    set(IFACE_IS_SPMD_DPC ON)
+else()
+    set(IFACE_IS_SPMD_DPC OFF)
+endif()
+
+if(IFACE_IS_SPMD_DPC)
     message(STATUS "MPI_INCLUDE_DIRS:" ${MPI_INCLUDE_DIRS})
     message(STATUS "MPI_LIBRARY_DIR:" ${MPI_LIBRARY_DIR})
 endif()
@@ -105,9 +117,9 @@ if(IFACE STREQUAL "host")
 
     list(APPEND COMPILE_DEFINITIONS "NPY_NO_DEPRECATED_API=NPY_1_7_API_VERSION")
 
-elseif(IFACE STREQUAL "dpc" OR IFACE STREQUAL "spmd_dpc")
+elseif(IFACE_IS_DPC OR IFACE_IS_SPMD_DPC)
 
-    if(IFACE STREQUAL "spmd_dpc")
+    if(IFACE_IS_SPMD_DPC)
         set(TARGET "_onedal_py_spmd_dpc")
     else()
         set(TARGET "_onedal_py_dpc")
@@ -150,7 +162,7 @@ elseif(IFACE STREQUAL "dpc" OR IFACE STREQUAL "spmd_dpc")
         endif()
     endif()
 
-    if(IFACE STREQUAL "spmd_dpc")
+    if(IFACE_IS_SPMD_DPC)
         set(MPI_LIBRARY ${MPI_LIBS})
     endif()
 
@@ -163,7 +175,7 @@ elseif(IFACE STREQUAL "dpc" OR IFACE STREQUAL "spmd_dpc")
              "ONEDAL_DPCTL_INTEGRATION")
     endif()
 
-    if(IFACE STREQUAL "spmd_dpc")
+    if(IFACE_IS_SPMD_DPC)
         list(APPEND COMPILE_DEFINITIONS
              "ONEDAL_DATA_PARALLEL_SPMD")
     endif()
@@ -195,7 +207,7 @@ if(ONEDAL_DPCTL_INTEGRATION)
 list(APPEND EXTERNAL_INCLUDE_DIRECTORIES ${DPCTL_INCLUDE_DIR})
 endif()
 
-if(IFACE STREQUAL "spmd_dpc")
+if(IFACE_IS_SPMD_DPC)
 list(APPEND EXTERNAL_LINK_LIBRARIES ${MPI_LIBRARY})
 list(APPEND EXTERNAL_LINK_DIRECTORIES ${MPI_LIBRARY_DIR})
 list(APPEND EXTERNAL_INCLUDE_DIRECTORIES ${MPI_INCLUDE_DIRS})
